Get nearest frontier centroid
Find the frontier centroid closest to the robot. In a future commit, the
robot will be set to move towards this closest centroid. Modify the
traverse_frontier function to use the 8-neighbors rather 4-neighbors
since frontiers are often diagonal and we want to consider diagonal
frontier cells as part of the same frontier when calculating centroids.